travel-meetup/requirements.txt
Django==4.2.7
Pillow==10.1.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7

